backend/pyproject.toml
[project]
name = "lavendersentinel"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "apscheduler>=3.11.2",
    "arxiv>=2.4.0",
    "litellm>=1.80.0",
    "marker>=2.1.3",
    "marker-pdf>=1.7.2",
    "psycopg2-binary>=2.9.11",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.12.0",
    "pypdf>=6.5.0",
    "pyyaml>=6.0.3",
    "sentence-transformers>=5.2.0",
    "sqlalchemy>=2.0.45",
    "streamlit-pdf-viewer>=0.0.26",
    "streamlit[pdf]>=1.52.2",
    "tqdm>=4.67.1",
    "redis>=5.0.0",
    "rq>=2.0.0",
    "supervisor>=4.2.0",
]
